pyproject.toml
[project]
name = "sub-agents"
version = "0.1.0"
description = "MCP server providing Gemini sub-agents for AI coding assistants"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["mcp", "gemini", "ai", "agent", "coding-assistant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "fastmcp>=2.0.0",
    "mcp>=1.0.0",
    "anyio>=4.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "psutil>=5.9.0",
]

[project.urls]
Homepage = "https://github.com/DoraemonHugU/sub-agents"
Repository = "https://github.com/DoraemonHugU/sub-agents"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
